Move Bullet-related code from OGFILE3.cpp into new file OGFILE_BULLETS.cpp.
