+++ src/OGFILE3.cpp
@@ -51,258 +51,6 @@ static char* create_rebel_func();
 static void write_ai_info(File* filePtr, short* aiInfoArray, short aiInfoCount, short aiInfoSize);
 static void read_ai_info(File* filePtr, short** aiInfoArrayPtr, short& aiInfoCount, short& aiInfoSize);
 
-//*****//
-
-
-//-------- Start of function BulletArray::write_file -------------//
-//
-int BulletArray::write_file(File* filePtr)
-{
-	filePtr->file_put_short(restart_recno);  // variable in SpriteArray
-
-	int    i, emptyRoomCount=0;;
-	Bullet *bulletPtr;
-
-	filePtr->file_put_short( size() );  // no. of bullets in bullet_array
-
-	for( i=1; i<=size() ; i++ )
-	{
-		bulletPtr = (Bullet*) get_ptr(i);
-
-		//----- write bulletId or 0 if the bullet is deleted -----//
-
-		if( !bulletPtr )    // the bullet is deleted
-		{
-			filePtr->file_put_short(0);
-			emptyRoomCount++;
-		}
-		else
-		{
-			filePtr->file_put_short(bulletPtr->sprite_id);      // there is a bullet in this record
-
-			//------ write data in the base class ------//
-
-			if( !bulletPtr->write_file(filePtr) )
-				return 0;
-
-			//------ write data in the derived class -------//
-
-			if( !bulletPtr->write_derived_file(filePtr) )
-				return 0;
-		}
-	}
-
-	//------- write empty room array --------//
-
-	write_empty_room(filePtr);
-
-	//------- verify the empty_room_array loading -----//
-
-#ifdef DEBUG
-	err_when( empty_room_count != emptyRoomCount );
-
-   for( i=0 ; i<empty_room_count ; i++ )
-   {
-		if( !is_deleted( empty_room_array[i].recno ) )
-         err_here();
-   }
-#endif
-
-	return 1;
-}
-//--------- End of function BulletArray::write_file -------------//
-
-
-//-------- Start of function BulletArray::read_file -------------//
-//
-int BulletArray::read_file(File* filePtr)
-{
-	restart_recno    = filePtr->file_get_short();
-
-	int     i, bulletRecno, bulletCount, emptyRoomCount=0, spriteId;
-	Bullet* bulletPtr;
-
-	bulletCount = filePtr->file_get_short();  // get no. of bullets from file
-
-	for( i=1 ; i<=bulletCount ; i++ )
-	{
-		spriteId = filePtr->file_get_short();
-		if( spriteId == 0 )
-		{
-			add_blank(1);     // it's a DynArrayB function
-
-			emptyRoomCount++;
-		}
-		else
-		{
-			//----- create bullet object -----------//
-
-			bulletRecno = create_bullet(spriteId);
-			bulletPtr   = bullet_array[bulletRecno];
-
-         //----- read data in base class --------//
-
-         if( !bulletPtr->read_file( filePtr ) )
-            return 0;
-
-			//----- read data in derived class -----//
-
-			if( !bulletPtr->read_derived_file( filePtr ) )
-				return 0;
-      }
-	}
-
-   //-------- linkout() those record added by add_blank() ----------//
-	//-- So they will be marked deleted in DynArrayB and can be -----//
-	//-- undeleted and used when a new record is going to be added --//
-
-	for( i=1 ; i<=size() ; i++ )
-	{
-		DynArrayB::go(i);             // since BulletArray has its own go() which will call GroupArray::go()
-
-		if( get_ptr() == NULL )       // add_blank() record
-			linkout();
-	}
-
-	//------- read empty room array --------//
-
-	read_empty_room(filePtr);
-
-	//------- verify the empty_room_array loading -----//
-
-#ifdef DEBUG
-	err_when( empty_room_count != emptyRoomCount );
-
-	for( i=0 ; i<empty_room_count ; i++ )
-	{
-		if( !is_deleted( empty_room_array[i].recno ) )
-			err_here();
-	}
-#endif
-
-	return 1;
-}
-//--------- End of function BulletArray::read_file ---------------//
-
-template <typename Visitor>
-static void visit_bullet(Visitor *v, Bullet *b)
-{
-	visit_sprite_members(v, b);
-	visit<int8_t>(v, &b->parent_type);
-	visit<int16_t>(v, &b->parent_recno);
-	visit<int8_t>(v, &b->target_mobile_type);
-	visit<float>(v, &b->attack_damage);
-	visit<int16_t>(v, &b->damage_radius);
-	visit<int16_t>(v, &b->nation_recno);
-	visit<int8_t>(v, &b->fire_radius);
-	visit<int16_t>(v, &b->origin_x);
-	visit<int16_t>(v, &b->origin_y);
-	visit<int16_t>(v, &b->target_x_loc);
-	visit<int16_t>(v, &b->target_y_loc);
-	visit<int8_t>(v, &b->cur_step);
-	visit<int8_t>(v, &b->total_step);
-}
-
-enum { BULLET_RECORD_SIZE = 57 };
-
-//--------- Begin of function Bullet::write_file ---------//
-//
-int Bullet::write_file(File* filePtr)
-{
-	return visit_with_record_size<FileWriterVisitor>(filePtr, this, &visit_bullet<FileWriterVisitor>,
-											BULLET_RECORD_SIZE);
-}
-//----------- End of function Bullet::write_file ---------//
-
-//--------- Begin of function Bullet::read_file ---------//
-//
-int Bullet::read_file(File* filePtr)
-{
-	if (!visit_with_record_size<FileReaderVisitor>(filePtr, this, &visit_bullet<FileReaderVisitor>,
-										BULLET_RECORD_SIZE))
-		return 0;
-
-   //------------ post-process the data read ----------//
-
-	sprite_info = sprite_res[sprite_id];
-
-	sprite_info->load_bitmap_res();
-
-	return 1;
-}
-//----------- End of function Bullet::read_file ---------//
-
-
-//----------- Begin of function Bullet::write_derived_file ---------//
-int Bullet::write_derived_file(File *filePtr)
-{
-	//--- write data in derived class -----//
-
-	int writeSize = bullet_array.bullet_class_size(sprite_id)-sizeof(Bullet);
-
-	if( writeSize > 0 )
-	{
-		if( !filePtr->file_write( (char*) this + sizeof(Bullet), writeSize ) )
-			return 0;
-	}
-
-	return 1;
-
-}
-//----------- End of function Bullet::write_derived_file ---------//
-
-
-//----------- Begin of function Bullet::read_derived_file ---------//
-int Bullet::read_derived_file(File *filePtr)
-{
-	//--- read data in derived class -----//
-
-	int readSize = bullet_array.bullet_class_size(sprite_id) - sizeof(Bullet);
-
-	if( readSize > 0 )
-	{
-		if( !filePtr->file_read( (char*) this + sizeof(Bullet), readSize ) )
-			return 0;
-	}
-
-	return 1;
-}
-//----------- End of function Bullet::read_derived_file ---------//
-
-template <typename Visitor>
-static void visit_projectile(Visitor *v, Projectile *p)
-{
-	visit<float>(v, &p->z_coff);
-	visit_sprite_members(v, &p->act_bullet);
-	visit_sprite_members(v, &p->bullet_shadow);
-}
-
-enum { PROJECTILE_RECORD_SIZE = 72 };
-
-//----------- Begin of function Projectile::read_derived_file ---------//
-
-int Projectile::write_derived_file(File *filePtr)
-{
-	return visit_with_record_size<FileWriterVisitor>(filePtr, this, &visit_projectile<FileWriterVisitor>,
-											PROJECTILE_RECORD_SIZE);
-}
-
-int Projectile::read_derived_file(File *filePtr)
-{
-	if (!visit_with_record_size<FileReaderVisitor>(filePtr, this, &visit_projectile<FileReaderVisitor>,
-										PROJECTILE_RECORD_SIZE))
-		return 0;
-
-   //----------- post-process the data read ----------//
-	act_bullet.sprite_info = sprite_res[act_bullet.sprite_id];
-	act_bullet.sprite_info->load_bitmap_res();
-	bullet_shadow.sprite_info = sprite_res[bullet_shadow.sprite_id];
-	bullet_shadow.sprite_info->load_bitmap_res();
-
-	return 1;
-}
-//----------- End of function Projectile::read_derived_file ---------//
-
 
 //*****//
